Address PR review: handshake scoping, teardown races, 64-bit page index
Tag allocation is now keyed by DistributedContextId and the ordered rank pair,
as rank-scoped MeshSockets are. A process-wide counter drifts between two ranks
that took part in different sockets before connecting to each other, after
which they pick different tags and block in the handshake.

Both endpoints now agree on transport availability before either constructs
one. The handshake is a blocking send/recv pair, so one rank throwing while the
other waits would hang rather than fail both.

Barriers run on a two-rank sub-context instead of the full context. Ranks that
are neither sender nor receiver return early and never call the barrier, so any
world larger than two deadlocked. create_sub_context uses MPI_Comm_create_group,
so ranks outside the group are not waited on.

Move assignment unregisters the destination's relays before replacing impl_.
Assigning over a live socket previously destroyed its Impl without running the
destructor, leaving RelayLoop polling freed socket and transport objects.

Relay error state is published safely: the flag is atomic so the barrier's hot
loop can test it without locking, and the message is copied under a lock because
the relay thread writes it. describe() now goes through RelayLoop, which holds
the same mutex poll_once() holds for a whole sweep -- formatting live counters
from the barrier thread was a data race.

first_page is 64-bit through the interface. The relay's absolute counter was
narrowed to uint32_t, and neither use site survives truncation: MPI compares it
against a uint64_t counter, where the narrow value is promoted rather than
wrap-compared, and the ring index is a modulo that only tolerates truncation
when num_pages divides 2^32. At 14336 B pages that is 61.6 TB per connection,
about eleven hours of streaming, not a theoretical limit.

Also: ttsl::pause() instead of __builtin_ia32_pause(), which did not compile on
AArch64; const-ref the DistributedContext handle in seven test sites; drop the
kMaxOutstandingBatches and kSlotsPerBatch constants the verbs removal orphaned.

+++ tt_metal/distributed/host_transport/host_transport.hpp
@@ -40,9 +40,11 @@ public:
 
     // Room for a batch of `pages` right now.
     virtual bool can_send(uint32_t pages) const = 0;
-    // Hand `pages` pages from ring page `first_page` to the peer, which places
-    // them at the same index. False if nothing was sent.
-    virtual bool send(uint32_t first_page, uint32_t pages) = 0;
+    // Hand `pages` pages from absolute ring page `first_page` to the peer, which
+    // places them at the same index. False if nothing was sent. `first_page` stays
+    // 64-bit: it is compared against an absolute counter and only narrowed after
+    // the modulo, so truncating here would break once a stream passes 2^32 pages.
+    virtual bool send(uint64_t first_page, uint32_t pages) = 0;
     // Source bytes the transport has finished reading, so the relay can release
     // them back to the device. Absolute.
     virtual uint64_t pages_released() const = 0;
